+++ packaging/nix/default.nix
@@ -14,7 +14,7 @@
 
 rustPlatform.buildRustPackage {
   pname = "visage";
-  version = "0.2.0";
+  version = "0.3.0";
 
   src = lib.cleanSource ../..;